+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 # Changelog
 
 Historical list of changes in releases
-## [v0.0.42] - 2024-11-07
+
+## [v0.0.44] - 2025-01-23
+* Add new `createWhiteboard` method, and update exist whiteboard methods ([#48058f61](https://github.com/digitalsamba/embedded-sdk/commit/48058f61a85a5f8095a4d8bf7eabd1b4be57ad02)).
+* Add new methods for managing library files ([#870ce040](https://github.com/digitalsamba/embedded-sdk/commit/870ce040e618131106595fd242f2f6b8db4b9cd8)).
+
+## [v0.0.43] - 2025-01-09
+* Add a `templateParams` param to init options ([#c8c5c43e](https://github.com/digitalsamba/embedded-sdk/commit/8f3fb6c678cd285afdf031990d9bdfb8ad3790fe)).
+
+## [v0.0.42] - 2024-12-27
 * Don't try to request user media to enumerate media devices if no devices were preconfigured ([#cce3487](https://github.com/digitalsamba/embedded-sdk/commit/cce34876ad17935c76d0d94f3472ff822a43bd72)).
+* Add a list of new methods for whiteboard management `openWhiteboard`, `closeWhiteboard`, `toggleWhiteboard` [#cd611f3](https://github.com/digitalsamba/embedded-sdk/commit/cd611f377a64ccaa47273b72bde82c05a3f39212)
+* Add `addImageToWhiteboard` method (#4694a47)[https://github.com/digitalsamba/embedded-sdk/commit/4694a478b960465af58f6c29f1ae1000afe8765c]
+* Add `speakerStoppedTalking` event to type definitions (#d0510bb)[https://github.com/digitalsamba/embedded-sdk/commit/d0510bbcad6d4cfb74a2d089af279ae84f866637]
 
 ## [v0.0.41] - 2024-09-17
 * Support `userLeftBatch` event from frontend, emit individual `userLeft` events for users in the payload. ([#6474ee5](https://github.com/digitalsamba/embedded-sdk/commit/6474ee5a38b1d557ade2ed6fcc9411eab0287f83)).